update150120201158

+++ 8_AccesFichier/TD1/Exercice4/biblio.h
@@ -26,7 +26,7 @@ typedef struct personne{
 	char nom[NBMAXCAR];
 	char prenom[NBMAXCAR];
 	int age;
-	float poids;
+	int poids;
 	char sexe ; // 'f' ou 'm'
 }typePersonne;
